move mostly to longs now

+++ trufflesqueak/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/SqueakImageContext.java
@@ -153,7 +153,7 @@ public class SqueakImageContext {
         return language;
     }
 
-    public SmallInteger wrapInt(int i) {
+    public SmallInteger wrapInt(long i) {
         return new SmallInteger(this, i);
     }
