+++ trufflesqueak/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/model/ImmediateCharacter.java
@@ -35,7 +35,7 @@ public class ImmediateCharacter extends ImmutableObject {
     }
 
     @Override
-    public int unwrapInt() {
+    public long unwrapInt() {
         return getValue();
     }
